templates/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.7.4
click==8.1.7
colorama==0.4.6
distlib==0.3.8
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.115.0
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mysql-connector-python==9.0.0
passlib==1.7.4
platformdirs==4.2.2
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.1.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.31
sqlparse==0.5.1
starlette==0.38.5
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.3
virtualenv==20.26.3
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.3
